third_party/WebKit/LayoutTests/netinfo/basic-operation-expected.txt

Tests the basic operation of NetInfo.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS navigator.connection is defined.
PASS navigator.connection.type is defined.
PASS navigator.connection.downlinkMax is defined.
PASS typeof connection.type is "string"
PASS connection.type is newConnectionType
PASS connection.downlinkMax is newDownlinkMax
PASS typeof connection.type is "string"
PASS connection.type is initialType
PASS connection.downlinkMax is initialDownlinkMax
PASS successfullyParsed is true

TEST COMPLETE
